requirements.txt
PyQt6
imageio
imageio-ffmpeg
numpy
Pillow
